requirements.txt
africastalking==2.0.2
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
dnspython==2.8.0
filelock==3.20.3
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
itemadapter==0.13.1
itemloaders==1.3.2
jmespath==1.0.1
lxml==6.0.2
packaging==25.0
parsel==1.10.0
pluggy==1.6.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
Pygments==2.19.2
pymongo==4.16.0
pyOpenSSL==25.3.0
pytest==9.0.2
python-dotenv==1.2.1
PyYAML==6.0.3
queuelib==1.8.0
requests==2.32.5
requests-file==3.0.1
responses==0.25.8
schema==0.7.8
Scrapy==2.14.1
service-identity==24.2.0
soupsieve==2.8.1
tldextract==5.3.1
Twisted==25.5.0
typing_extensions==4.15.0
urllib3==2.6.3
w3lib==2.3.1
zope.interface==8.2
